+++ python/sglang/srt/debug_utils/schedule_simulator/simulator.py
@@ -1,22 +1,11 @@
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from typing import Any, Dict, List, Optional
 
-from sglang.srt.debug_utils.schedule_simulator.gpu_state import GPUState
+from sglang.srt.debug_utils.schedule_simulator.gpu_state import GPUState, StepRecord
 from sglang.srt.debug_utils.schedule_simulator.metrics import MetricRecorder
-from sglang.srt.debug_utils.schedule_simulator.request import RequestStage, SimRequest
+from sglang.srt.debug_utils.schedule_simulator.request import SimRequest
 from sglang.srt.debug_utils.schedule_simulator.routers.base import RouterPolicy
 from sglang.srt.debug_utils.schedule_simulator.schedulers.base import SchedulerPolicy
-
-
-@dataclass
-class StepRecord:
-    step: int
-    gpu_id: int
-    running_count: int
-    pending_count: int
-    total_seq_len: int
-    running_req_ids: List[str] = field(default_factory=list)
-    pending_req_ids: List[str] = field(default_factory=list)
 
 
 @dataclass
@@ -58,7 +47,7 @@ class Simulator:
             incoming_requests.clear()
             self._schedule_all_gpus()
             self._execute_step()
-            self._collect_step_records(step_records)
+            step_records.extend(gpu.get_step_record(self.step) for gpu in self.gpu_states)
             self._log_step()
             self._record_metrics()
             self.step += 1
@@ -85,29 +74,7 @@ class Simulator:
 
     def _execute_step(self) -> None:
         for gpu in self.gpu_states:
-            finished = []
-            for req in gpu.running_requests:
-                if req.stage == RequestStage.PREFILL:
-                    req.stage = RequestStage.DECODE
-                req.decoded_tokens += 1
-                if req.is_finished():
-                    finished.append(req)
-            for req in finished:
-                gpu.running_requests.remove(req)
-
-    def _collect_step_records(self, step_records: List[StepRecord]) -> None:
-        for gpu in self.gpu_states:
-            step_records.append(
-                StepRecord(
-                    step=self.step,
-                    gpu_id=gpu.gpu_id,
-                    running_count=len(gpu.running_requests),
-                    pending_count=len(gpu.pending_requests),
-                    total_seq_len=gpu.total_seq_len(),
-                    running_req_ids=[r.request_id for r in gpu.running_requests],
-                    pending_req_ids=[r.request_id for r in gpu.pending_requests],
-                )
-            )
+            gpu.execute_step()
 
     def _log_step(self) -> None:
         if self.log_level == 0:
